Pass the right priorXMax to ValidateTuplesXact()
In zheap_get_latest_tid(), ZHeapTupleGetTransInfo() was getting called
with 'resulttup'. But because 'resulttup' is returned from
ZHeapTupleSatisfiesVisibility(), it can be a tuple generated from an
undo record. And if we use ZHeapTupleGetTransInfo() to get the xid
that modified resulttup, it returns the xid that created the original
tuple instead of the xid that modified the tuple. This results in
the wrong xid being passed as priorXMax to ValidateTuplesXact(), which
in turn leads to assertion failures.

So move the ZHeapTupleGetTransInfo() call before
ZHeapTupleSatisfiesVisibility() call, so that we could use 'tp' rather
than resulttup. 'tp' gets freed by ZHeapTupleSatisfiesVisibility(), so
we could not use tp after ZHeapTupleSatisfiesVisibility() call.

Added a new isolation test.