Bug fix in execution undo action
Set the proper slot on the TPD offset map while replaying undo actions
if the prior version of the tuple is pointing to the TPD slot.  Prior
to this we are simply overwriting the older tuple version but the slot
is not updated in the TPD offset map.

Dilip Kumar review by Amit Kapila