Allow freezing and reusing of the TPD slots
Currently, for zheap pages we allow to freeze slots of the
the all visible xids and also allow to reuse the slot of the
committed xids.

This patch is implementing the same for the TPD slots. The mechanism
of the freezing and reusing the tpd slots are same as for the
zheap page slots.

Dilip Kumar reviewed by Amit Kapila